refactor: separating the selectimage functionality into a component
